pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-agent"
version = "0.1.0"
description = "RAG-powered agent for SEC financial filings"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "numpy",
    "faiss-cpu",
    "google-genai",
    "pyyaml",
    "python-dotenv",
]

[project.optional-dependencies]
dev = ["pytest"]
pipeline = [
    "sec-api",
    "beautifulsoup4",
    "spacy",
    "tiktoken",
    "yfinance",
    "pandas",
    "google-cloud-aiplatform",
]

[tool.setuptools.packages.find]
where = ["src"]
